Refactor external review normalization and classifier
